warnings.
svn:sync-xref-src-repo-rev: r80326

+++ src/VBox/Frontends/VirtualBox/src/extensions/QIAdvancedToolBar.cpp
@@ -1,4 +1,4 @@
-/* $Id: QIAdvancedToolBar.cpp 42526 2012-08-02 10:31:28Z sergey.dubov@oracle.com $ */
+/* $Id: QIAdvancedToolBar.cpp 42994 2012-08-27 10:01:39Z ramshankar.venkataraman@oracle.com $ */
 /** @file
  *
  * VBox frontends: Qt GUI ("VirtualBox"):
@@ -40,7 +40,7 @@ public:
     QSize sizeHint() const
     {
         /* Make the button a little bit taller: */
-        return QSize(QToolButton::sizeHint().width(), QToolButton::sizeHint().height() * 1.2);
+        return QSize((int)QToolButton::sizeHint().width(), (int)QToolButton::sizeHint().height() * 1.2);
     }
 
     /* Set the button to ignore mouse-press events if checked: */
